backend/requirements.txt
asgiref==3.9.1
asttokens==3.0.1
attrs==25.3.0
beautifulsoup4==4.14.3
blessings==1.7
bqplot==0.12.45
branca==0.8.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
colour==0.1.5
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
earthengine-api==1.7.10
EditorConfig==0.17.1
ee==0.2
ee_extra==2025.7.2
eerepr==0.1.2
executing==2.2.1
ffmpeg-python==0.2.0
filelock==3.20.3
folium==0.20.0
fonttools==4.61.1
future==1.0.0
gdown==5.2.1
geeadd==2.1.0
geemap==0.19.1
geocoder==1.38.1
geojson==3.2.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.10
inflection==0.5.1
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipytree==0.2.2
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jsbeautifier==1.15.4
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
numpy==2.4.1
packaging==26.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
plotly==5.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyCRS==1.0.2
Pygments==2.19.2
pyparsing==3.3.2
pyperclip==1.11.0
pyshp==3.0.3
PySocks==1.7.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.36.2
requests==2.32.4
rich==14.3.1
rpds-py==0.27.0
rsa==4.9.1
sankee==0.3.0
scooby==0.11.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.3
typing_extensions==4.14.1
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.4.0
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.15
xyzservices==2025.11.0
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0  